% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom.lsn.plot.R
\name{chrom.lsn.plot}
\alias{chrom.lsn.plot}
\title{Chromosome Lesion Plot}
\usage{
chrom.lsn.plot(
  grin.res,
  genome,
  lsn.clrs = NULL,
  chrom = NULL,
  plot.start = NULL,
  plot.end = NULL,
  lesion.grp = NULL,
  spec.lsn.clr = NULL,
  expand = 5e-04,
  hg38.transcripts = NULL,
  hg19.cytoband = NULL,
  hg38.cytoband = NULL
)
}
\arguments{
\item{grin.res}{GRIN results (output of the grin.stats function).}

\item{genome}{Function support either "hg19" or "hg38" genome assemblies based on the genome assembly that has been used to prepare the lesion data.}

\item{lsn.clrs}{Lesion colors for the regional gene plot (If not provided by the user, colors will be automatically assigned using default.grin.colors function).}

\item{chrom}{chromosome number}

\item{plot.start}{Start position of the locus of interest.}

\item{plot.end}{End position of the locus of interest.}

\item{lesion.grp}{Lesion type of interest (if not specified the plot will return all lesion groups)}

\item{spec.lsn.clr}{Assigned color for the lesion type of interest (should be specified when lesion.grp is specified).}

\item{expand}{Controls ratio of the specified locus (start and end position) to the whole plot with default value = 0.0005 (setting expand=0 will only plot the specified locus from the start to the end position without any of the upstream or downstream regions.}

\item{hg38.transcripts}{transcripts data retrieved from annotation hub for hg38 version 110 (should be only specified if genome="hg38").}

\item{hg19.cytoband}{hg19 chromosome bands start and end data in base pair (should be only specified if genome="hg19").}

\item{hg38.cytoband}{hg38 chromosome bands start and end data in base pair (should be only specified if genome="hg38").}
}
\value{
Function will return a plot of a pre-specfied lesion type or all lesion affecting a specific locus or the whole chromosome.
}
\description{
Function plot lesion data of a selected lesion type or all lesion groups located on a certain region such as a chromosome band or the whole specified chromosome .
}
\details{
Function will return a plot with either lesions of a certain lesion type specified by the user or all lesions of different lesion groups allocated to either a specific part of the chromosome such as a chromosome band or the whole length of the chromosme.As compared the the grin.gene.plot function, chrom.lsn.plot will not return a transcripts track which will allow the user to specify a large region or even the whole chromosome. A chromosome ideogram is added on the top of the lesion plot with the specified part of the chromosome encircled in red.
}
\examples{
\dontrun{
data(lesion.data)
data(hg19.gene.annotation)
data(hg19.chrom.size)
data(hg19_cytoband)
data(hg38_cytoband)

# run GRIN analysis using grin.stats function
grin.results=grin.stats(lesion.data,
                        genome.version="Human_GRCh37")

# Plots Showing Different Types of Lesions Affecting a region of Interest without plotting the
# transcripts track (this will allow plotting a larger locus of the chromosome such as a
# chromosome band using hg19 genome assembly):
locus.plot=chrom.lsn.plot(grin.results, genome="hg19", hg19.cytoband=hg19_cytoband,
                          chrom=9, plot.start=19900000, plot.end=25600000,
                          lesion.grp = "loss", spec.lsn.clr = "blue")

# Plots Showing Different Types of Lesions Affecting the whole chromosome:
chrom.lsn=chrom.lsn.plot(GRIN.results, genome="hg19", hg19.cytoband=hg19_cytoband,
                          chrom=9, plot.start=1, plot.end=141000000)

# for hg38 genome assembly:
ah <- AnnotationHub()
gtf.V110 <- ah[["AH113665"]]

# run GRIN analysis using grin.stats function
grin.results=grin.stats(lesion.data,
                       genome.version="Human_GRCh38")
# to plot deletions on the p21.3 band of chromosome 9:
chrom.lsn.plot(grin.results, genome="hg38", hg38.transcripts="gtf.v110",
               hg38.cytoband=hg38_cytoband, chrom=9, plot.start=19900000, plot.end=25600000,
               lesion.grp = "loss", spec.lsn.clr = "blue")

# to plot all lesions on chromosome 9:
chrom.lsn.plot(grin.results, genome="hg38", hg38.transcripts="gtf.v110",
               hg38.cytoband=hg38_cytoband, chrom=9, plot.start=1, plot.end=141000000)
}
}
\references{
Cao, X., Elsayed, A. H., & Pounds, S. B. (2023). Statistical Methods Inspired by Challenges in Pediatric Cancer Multi-omics.
}
\seealso{
\code{\link[=grin.stats]{grin.stats()}}
}
\author{
{Abdelrahman Elsayed \email{abdelrahman.elsayed@stjude.org} and Stanley Pounds \email{stanley.pounds@stjude.org}}
}
