% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.data.R
\docType{data}
\name{expr.data}
\alias{expr.data}
\title{Example T-ALL Dataset Gene Expression Data}
\format{
\subsection{\code{expr.data}}{

A data frame with 420 rows and 265 columns:
\describe{
\item{gene}{Ensembl IDs of the list of 420 genes included in the dataset}
...
}
}
}
\source{
Data was extracted from the supplementary material tables of the published Liu, Yu, et al. (2017) manuscript \url{https://www.nature.com/articles/ng.3909#Sec27}
}
\usage{
expr.data
}
\description{
Gene expression data file showing log2 normalized expression level of 420 genes (rows) in 265 newly diagnosed T-cell Acute Lymphoblastic Leukemia (T-ALL) patients in columns that was reported by Liu, Yu, et al. (2017).
}
\keyword{datasets}
