% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg19_cytoband.R
\docType{data}
\name{hg19_cytoband}
\alias{hg19_cytoband}
\title{GRCh37 Chromosome Cytobands}
\format{
\subsection{\code{hg19_cytoband}}{

A data frame with 862 rows and 5 columns:
\describe{
\item{chrom}{The chromosome number.}
\item{chromStart}{The cytoband start position on the chromosome in base pairs.}
\item{chromEnd}{The cytoband end position on the chromosome in base pairs.}
\item{name}{The cytoband name.}
\item{gieStain}{The coloring scheme of the cytobands.}
}
}
}
\source{
The Chromosome cytobands data file was downloaded from the UCSC genome browser for GRCh37 genome assembly \url{https://hgdownload.soe.ucsc.edu/goldenPath/hg19/database/}.
}
\usage{
hg19_cytoband
}
\description{
The dataset has the start and end positions in base pairs of all 22 autosomes in addition to X and Y chromosome cytobands for Human-GRCh37 (hg19) genome assembly.
}
\keyword{datasets}
