% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg38_chrom_size.R
\docType{data}
\name{hg38_chrom_size}
\alias{hg38_chrom_size}
\title{Chromosome Length Data (hg38)}
\format{
\subsection{\code{hg38_chrom_size}}{

A data frame with 24 rows and 2 columns:
\describe{
\item{chrom}{Chromosome name (1,2, 3, ..., X, Y).}
\item{size}{Chromosome length in base pairs.}
}
}
}
\source{
Retrieved from UCSC Genome Browser \code{chr.info} text files using the \code{get.chrom.length} function with the GRCh38 genome assembly.
}
\usage{
hg38_chrom_size
}
\description{
This dataset contains the lengths (in base pairs) of the 22 autosomes in addition to the X and Y chromosomes based on the GRCh38 human genome assembly. The data was retrieved from the UCSC Genome Browser using the \code{get.chrom.length} function with "Human-GRCh38" as the selected genome.
}
\keyword{datasets}
