% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.index.gene.data.R
\name{order.index.gene.data}
\alias{order.index.gene.data}
\title{Order Index Gene Data}
\usage{
order.index.gene.data(gene.data)
}
\arguments{
\item{gene.data}{data.frame with gene annotation data either provided by the user or retrieved from ensembl BioMart database using get.ensembl.annotation function included in the GRIN2.0 library. data.frame should has four columns that include "gene" which is the ensembl ID of the annotated genes to which the lesion data will be overlapped, "chrom" which is the chromosome on which the gene is located, "loc.start" which is the gene start position, and "loc.end" the gene end position.}
}
\value{
A list with the following components:
\item{gene.data}{Input gene annotation data}
\item{gene.index}{data.frame with two columns of ordered row start and row end based on the number of genes annotated to each chromosome.}
}
\description{
This function order and index gene annotation data by chromosome on which the gene is located, gene start, and end positions.
}
\examples{
data(hg19.gene.annotation)

ordered.genes=order.index.gene.data(hg19.gene.annotation)
}
\references{
Pounds, Stan, et al. (2013) A genomic random interval model for statistical analysis of genomic lesion data.

Cao, X., Elsayed, A. H., & Pounds, S. B. (2023). Statistical Methods Inspired by Challenges in Pediatric Cancer Multi-omics.
}
\author{
{Stanley Pounds \email{stanley.pounds@stjude.org}}
}
