% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.R
\docType{data}
\name{pathways}
\alias{pathways}
\title{Genes Annotated to Biological Pathways}
\format{
\subsection{\code{pathways}}{

A data frame with 121 rows and 3 columns:
\describe{
\item{gene.name}{Gene symbol.}
\item{ensembl.id}{Ensembl gene ID.}
\item{pathway}{Biological pathway to which the gene is annotated.}
}
}
}
\source{
Extracted from the supplementary materials of Liu, Yu, et al. (2017)
\url{https://www.nature.com/articles/ng.3909#Sec27}
}
\usage{
pathways
}
\description{
A dataset listing genes annotated to various biological pathways, as reported in Liu, Yu, et al. (2017).
}
\keyword{datasets}
