% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg38_cytoband.R
\docType{data}
\name{hg38_cytoband}
\alias{hg38_cytoband}
\title{GRCh38 Chromosome Cytobands}
\format{
\subsection{\code{hg38_cytoband}}{

A data frame with 1,549 rows and 5 columns:
\describe{
\item{chrom}{The chromosome number.}
\item{chromStart}{The cytoband start position on the chromosome in base pairs.}
\item{chromEnd}{The cytoband end position on the chromosome in base pairs.}
\item{name}{The cytoband name.}
\item{gieStain}{The coloring scheme of the cytobands.}
}
}
}
\source{
The Chromosome cytobands data file was downloaded from the UCSC genome browser for GRCh38 genome assembly \url{https://hgdownload.soe.ucsc.edu/goldenPath/hg38/database/}.
}
\usage{
hg38_cytoband
}
\description{
The dataset has the start and end positions in base pairs of all 22 autosomes in addition to X and Y chromosome cytobands for Human-GRCh38 (hg38) genome assembly.
}
\keyword{datasets}
